+++ RELEASES.md
@@ -26,6 +26,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: allow for null declaration in bindings
 * prov-template-compiler: compiler support for QualifiedSpecializationOf
 * prov-template-compiler: CompilerConfigurations: generate Relation0Configurator
+* prov-template-compiler: CompilerConfigurations: generate RelationConfigurator
 
 
 `2.2.1`